requirements.txt
fastapi[all]==0.112.0
pydantic==2.8.2
numpy==1.26.4
onnxruntime==1.21.0
Pillow==11.1.0
jinja2==3.1.2